// core/web-app/src/types/bindings/cloudwatch-statistics-metric.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface CloudwatchStatisticsMetricMetadata { polling_interval: number, access_key: string, secret_key: string, region: string, namespace: string, metric_name: string, dimensions: Array<string>, statistic: string, period: number, unit: string, duration_seconds: number, }